add: very small refactoring

+++ server/src/handlers/del.cpp
@@ -10,14 +10,16 @@ using json = nlohmann::json;
 std::string HandlerOn::del(const HttpRequestStruct& http) {
 	DB db(PATH_TO_DB);
 
-	std::string id = http.path.params.at("id") != "" ? http.path.params.at("id") : "";
+	if (http.path.params.find("id") == http.path.params.end())
+		return Http::response(400, "Missing 'id'");
+
+	std::string id = http.path.params.at("id");
 	std::string table = http.path.type;
 	DBResponseStruct response = db.delete_data(table, id);
 
 	std::string body = json(response.data).dump();
-	StatusStruct status = response.status;
 
-	return Http::response(status.code, status.msg, body, {
+	return Http::response(response.status, body, {
 		{ "Content-Type", "application/json" }
 	});
 }
